+++ firecore/import_utils.py
@@ -1,8 +1,9 @@
 from typing import Union, Any
 import importlib
-from firecore.logging import get_logger
+import logging
 
-logger = get_logger(__name__)
+
+logger = logging.getLogger(__name__)
 
 
 def require(name: str):