+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Use default rules `None` in `adapt`
+- Use std logging
 
 ## [0.5.0] - 2022-01-25