+++ CHANGELOG.md
@@ -7,10 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+## [0.6.0]
+
 ### Changed
 
 - Use default rules `None` in `adapt`
 - Use std logging
+- Introduce `naming` and `extract` to `adapter`.
+- Temporarily remove config system. I don't know how to config.
 
 ## [0.5.0] - 2022-01-25